It works!!!!
Got a slow pulse and I can adjust the duty cycle with the knob. Only thing is I
can't get a truly 0 duty cycle consistently (I can get stuck on ON). I'll fix
it. First, I want to try to remove the dependency on two timers (what a waste!)
by using the phase-corrected PWM on timer 1.